moved user to schemas removed token auth example
