+++ docker-compose_dev.yml
@@ -27,6 +27,7 @@ services:
       - auth_data:/usr/src/app/writable
     ports:
       - '3010:3010'
+    # todo typeorm url + config
     environment:
       PORT: 3010
       ADMIN_USER: admin
